crawler/crawler_auth/requirements.txt

 bs4==0.0.1
 celery==4.4.0
 channels==2.4.0
 django-countries==6.1.3
 django-xhtml2pdf==0.0.4
 djongo==1.3.3
 geopy==2.0.0
 lxml==4.5.2
 nest-asyncio==1.4.1
 Pillow==5.1.0
 pymongo==3.11.0
 twint==2.1.20
 xhtml2pdf==0.2.4
mongoengine==0.18.2
#django-background-tasks==1.2.0
#django-compat==1.0.15
#django-cors-headers==3.2.0
#django-eventstream==2.6.0
#django-grip==2.0.2
#djangorestframework==3.10.3

